% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Silent read of csv file into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Either a path to a file, a connection, or literal data (either a single string or a raw vector).}
}
\value{
This function returns a tibble object, exported from the user's csv file.
}
\description{
This function allows for reading a comma separated values and transforms them
into tibble objects. After checking that the file exists, the function suppress
messages produced by read_csv().
}
\details{
The function will not work correctly if provided with a wrong filepath or if a
file is not a csv object.
}
\examples{
fars_read(readr_example("mtcars.csv"))
my_df <- fars_read(readr_example("mtcars.csv"))

}
