% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading month data from the FARS file for a given year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{an R object. Must be an integer or a character vector of length 1, containing any year}
}
\value{
This function returns a tibble object, containing a column with month exported from the
data and a year for which the month data was accessed. If provided with miltiple years -- it
will return a list with data frames inside.
}
\description{
This function allows for a simplier reading of the data from the US Fatality Analysis
Reporting System. It will return two columns - column with a month and column containing a
requested year.
}
\details{
The function will give a warning if provided with an incorrect year. Note that
the function doesn't stop running if it found an error -- only produces a warning.
}
\examples{
my_df <- fars_read_years(2013)

}
