% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Calculate the number of observations per month for the FARS data file for a given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{an R object. Must be an integer or a character vector of length 1 or more,
containing any year}
}
\value{
This function returns a tibble object containing a column with month exported from the
data and a column with the number of observations per month. If provided with multiple years,
the data frame will also contain multiple columns with numbers of observations -- each for a year
}
\description{
For a given year this function allows for calculating the number of observations per
month for the US Fatality Analysis Reporting System data.
}
\details{
The function will give a warning if provided with an incorrect year. Note that
the function doesn't stop running if it found an error -- only produces an erorr.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))

}
