% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a unified filename for a compressed bz2 file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an R object. Must be an integer or a character vector of length 1, containing any year}
}
\value{
This function returns a character object
}
\description{
This function allows creating a filename using a year as an input. This function is
aimed to make reading files faster for the analysis of the data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System.
}
\details{
The function will provide NAs in the filename if provided with an object non-coercible to
an integer. The function will give errors if provided with an object with  a length longer than 1.
}
\examples{
make_filename(2013)
my_df <- fars_read(make_filename(2013))

}
